Only test on jvm
